refactor: extract bp_components to routes/components.py
Phase 5.9 of incremental modularisation. Moves the 5 component-panel detail
endpoints (tool / runtime / machine / gateway / brain) out of the 28k-line
dashboard.py monolith into routes/components.py. Pure mechanical move —
zero behaviour change.

- dashboard.py: 28074 -> 27072 lines (-1002)
- routes/components.py: new, 1040 lines

Module-level helpers (SESSIONS_DIR, LOG_DIR, _get_log_dirs, _grep_log_file,
_record_heartbeat, _provider_from_model, _get_sessions, _get_crons,
_get_memory_files, get_local_ip) stay in dashboard.py and are accessed via
late `import dashboard as _d` to avoid circular imports. The tool-response
cache (_api_tool_cache) moves with its sole consumer.
